+++ src/lib/aggregations/get-top-hits.spec.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-underscore-dangle */
-import { ResponseError } from '@elastic/elasticsearch/lib/errors.js'
+import { ResponseError } from 'lib/common'
 import { Order } from 'lib/enums'
 import { HomeDocument } from 'test/module'
 import { TEST_ELASTICSEARCH_NODE } from 'test/constants'
@@ -212,7 +212,8 @@ describe('getTopHitsAggregation', () => {
             })
             .catch(error => {
                 expect(error).toBeInstanceOf(ResponseError)
-                expect(error.message).toContain('search_phase_execution_exception: [illegal_argument_exception]')
+                expect(error.message).toContain('search_phase_execution_exception')
+                expect(error.message).toContain('illegal_argument_exception')
                 expect(error.message).toContain(
                     `Top hits result window is too large, the top hits aggregator [result]'s from + size must be less than or equal to: [100] but was [1000].`
                 )
@@ -233,7 +234,8 @@ describe('getTopHitsAggregation', () => {
             })
             .catch(error => {
                 expect(error).toBeInstanceOf(ResponseError)
-                expect(error.message).toContain('search_phase_execution_exception: [illegal_argument_exception]')
+                expect(error.message).toContain('search_phase_execution_exception')
+                expect(error.message).toContain('illegal_argument_exception')
                 expect(error.message).toContain(
                     `Top hits result window is too large, the top hits aggregator [result]'s from + size must be less than or equal to: [100] but was [107].`
                 )