+++ src/lib/common/errors.ts
@@ -1,0 +1,3 @@
+import { errors } from '@elastic/elasticsearch'
+
+export const ResponseError = errors.ResponseError