+++ src/lib/responses/search.ts
@@ -1,5 +1,5 @@
-import type { ApiResponse } from '@elastic/elasticsearch'
-import { ClassConstructor, Document, Result } from 'lib/common'
+import { SearchResponse as BaseSearchResponse } from '@elastic/elasticsearch/lib/api/types'
+import { ClassConstructor, Document } from 'lib/common'
 import { AggregationsContainer } from 'lib/aggregations'
 import { TDocumentWrapper, TransformedAggregations, getTransformedAggregations, getTransformedDocuments, getTransformedTotal } from 'lib/transformers'
 
@@ -11,9 +11,9 @@ export type SearchResponse<TDocument extends Document, TAggregationsBody extends
 
 export const getSearchResponse = <TDocument extends Document, TAggregationsBody extends AggregationsContainer<TDocument>>(
     document: ClassConstructor<TDocument>,
-    { body }: ApiResponse<Result<TDocument, TAggregationsBody>>
+    { hits, aggregations }: BaseSearchResponse<TDocument, TAggregationsBody>
 ): SearchResponse<TDocument, TAggregationsBody> => ({
-    total: getTransformedTotal(body.hits),
-    documents: getTransformedDocuments(document, body.hits),
-    aggregations: getTransformedAggregations(body.aggregations)
+    total: getTransformedTotal(hits),
+    documents: getTransformedDocuments(document, hits),
+    aggregations: getTransformedAggregations(aggregations)
 })