+++ src/lib/utils/is-total-hits.ts
@@ -1,8 +1,8 @@
-import { estypes } from '@elastic/elasticsearch'
 import { is } from 'ramda'
+import { SearchTotalHits } from '@elastic/elasticsearch/lib/api/types'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const isTotalHits = (object: any): object is estypes.TotalHits => {
+export const isTotalHits = (object: any): object is SearchTotalHits => {
     if (is(Object, object) && is(Number, object.value)) {
         return true
     }